Refactor UserCard into UserCardFront. Refactor UserLinks into UserCardBack. Replace UserLinks prop "links" with "user".
