+++ src/components/UserCard/UserCardFront/TypeIcon/TypeIcon.css
