+++ src/components/UserCard/UserCardFront/TypeIcon/TypeIcon.jsx
